Split tests after rebase
